Fix minor styling issue in widget blocks. Update text for open/close banner.
