+++ lib/shortcode_functions.php
@@ -1,0 +1,54 @@
+<?php
+
+require_once (realpath(dirname(__FILE__) . '/..') . '/blocks/blocks_autoload.php');
+
+use Blocks\Constants;
+
+function get_event_desc_shortcode(): string
+{
+    $hide = get_post_meta( get_the_ID(), Constants::CUSTOM_FIELD_HIDE_EVENT_TEXT, true );
+    $id = get_post_meta( get_the_ID(), Constants::CUSTOM_FIELD_TEMPLATE, true );
+    $desc = get_post_meta( get_the_ID(), Constants::CUSTOM_FIELD_EVENT_DESC, true );
+
+    $hide = hideCustomEventText($hide);
+    $id = trim($id);
+
+    //if there is no templateId or we set hide descriptions to "y" then return empty string
+    if(empty($id) || ($hide)) {
+        return '';
+    }
+
+    return "<div data-shortcode='event-desc' data-template-id='{$id}-desc'>$desc</div>";
+}
+
+function get_event_book_now_shortcode(): string
+{
+    $url = get_post_meta( get_the_ID(), Constants::CUSTOM_FIELD_BOOK_NOW, true );
+
+    $url = trim($url);
+
+    //if there is no templateId or we set hide descriptions to "y" then return empty string
+    if(empty($url)) {
+        return '';
+    }
+
+    return "<div class='is-layout-flex wp-block-buttons'>\n
+            <div class='wp-block-button'><p><a class='wp-block-button__link wp-element-button' href='$url'>BOOK NOW</a></p></div>\n
+            </div>";
+}
+
+function get_museum_hours_shortcode(): string
+{
+    $open = '11am';
+    $close = '5pm';
+    $open_today = false;
+    $next_open_day = 'Wednesday';
+
+    if ($open_today) {
+        $message =  "The Seaport Museum is open today {$open}–{$close}.";
+    } else {
+        $message = "The Seaport Museum will open $next_open_day {$open}–{$close}.";
+    }
+
+    return "<div data-shortcode='museum-hours'><span class='dashicons dashicons-clock'></span>$message</div>";
+}
